--- a/README.md
+++ b/README.md
@@ -45,6 +45,7 @@ For more information:
 
 AcceptanceTest logs are output to `bin/Debug/net6.0/log`.
 
+
 Credits
 -------