--- a/RELEASE_NOTES.md
+++ b/RELEASE_NOTES.md
@@ -11,6 +11,11 @@ What's New
 **CAUTION: There are breaking changes between 1.10 and 1.11!  Please review the 1.11.0 notes below.**
 
 ### NEXT RELEASE
+
+**Breaking change**
+* #768 - span-ify parser (Rob-Hague) - makes a change to QuickFix.Parser interface, which isn't likely to affect users
+
+**Non-breaking changes**
 * #400 - added DDTool, a C#-based codegen, and deleted Ruby-based generator (gbirchmeier)
 * #811 - convert AT platform to be NUnit-based, get rid of Ruby runner (Rob-Hague)
 * #813 - fix incorrect logging of acceptor heartbeat (gbirchmeier)