--- a/UnitTests/ParserTest.cs
+++ b/UnitTests/ParserTest.cs
@@ -1,6 +1,8 @@
 ﻿using NUnit.Framework;
 using QuickFix;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace UnitTests
 {
@@ -10,6 +12,7 @@ namespace UnitTests
         const string normalLength   = "8=FIX.4.2\x01" + "9=12\x01" + "35=A\x01" + " 108=30\x01" + "10=31\x01";
         const string badLength      = "8=FIX.4.2\x01" + "9=A\x01"  + "35=A\x01" + "108=30\x01"  + "10=31\x01";
         const string negativeLength = "8=FIX.4.2\x01" + "9=-1\x01" + "35=A\x01" + "108=30\x01"  + "10=31\x01";
+        const string zeroLength     = "8=FIX.4.2\x01" + "9=0\x01" + "35=A\x01" + "108=30\x01"  + "10=31\x01";
         const string incomplete_1   = "8=FIX.4.2";
         const string incomplete_2   = "8=FIX.4.2\x01" + "9=12";
 
@@ -29,64 +32,92 @@ namespace UnitTests
             Assert.AreEqual(12, len);
             Assert.AreEqual(15, pos);
 
-            pos = 0;
+            Assert.True(parser.ExtractLength(out len, out pos, zeroLength));
+            Assert.AreEqual(0, len);
+            Assert.AreEqual(14, pos);
+
             Assert.Throws<QuickFix.MessageParseError>(delegate { parser.ExtractLength(out len, out pos, badLength); });
-
             Assert.AreEqual(0, pos);
+
             Assert.Throws<QuickFix.MessageParseError>(delegate { parser.ExtractLength(out len, out pos, negativeLength); });
-
             Assert.AreEqual(0, pos);
-            parser.ExtractLength(out len, out pos, incomplete_1);
 
-            parser.ExtractLength(out len, out pos, incomplete_2);
+            Assert.False(parser.ExtractLength(out len, out pos, incomplete_1));
+            Assert.AreEqual(0, pos);
+
+            Assert.False(parser.ExtractLength(out len, out pos, incomplete_2));
             Assert.AreEqual(0, pos);
 
             Assert.False(parser.ExtractLength(out len, out pos, ""));
+            Assert.AreEqual(0, pos);
         }
 
         [Test]
-        public void ReadCompleteFixMessages()
+        [TestCase(100, 1)]
+        [TestCase(10, 10)]
+        public void ReadCompleteFixMessages(int batchSize, int numBatches)
         {
             const string fixMsg1 = "8=FIX.4.2\x01" + "9=12\x01" + "35=A\x01" + "108=30\x01" + "10=31\x01";
             const string fixMsg2 = "8=FIX.4.2\x01" + "9=17\x01" + "35=4\x01" + "36=88\x01"  + "123=Y\x01"  + "10=34\x01";
             const string fixMsg3 = "8=FIX.4.2\x01" + "9=19\x01" + "35=A\x01" + "108=30\x01" + "9710=8\x01" + "10=31\x01";
 
             Parser parser = new Parser();
-            byte[] combined = StrToBytes(fixMsg1 + fixMsg2 + fixMsg3);
-            parser.AddToStream(combined, combined.Length);
 
-            string readFixMsg1;
-            Assert.True(parser.ReadFixMessage(out readFixMsg1));
-            Assert.AreEqual(fixMsg1, readFixMsg1);
+            for (int batchNum = 0; batchNum < numBatches; batchNum++)
+            {
+                List<string> batch = new();
 
-            string readFixMsg2;
-            Assert.True(parser.ReadFixMessage(out readFixMsg2));
-            Assert.AreEqual(fixMsg2, readFixMsg2);
+                for (int i = 0; i < batchSize; i++)
+                {
+                    string message = (i % 3) switch
+                    {
+                        0 => fixMsg1,
+                        1 => fixMsg2,
+                        _ => fixMsg3
+                    };
 
-            string readFixMsg3;
-            Assert.True(parser.ReadFixMessage(out readFixMsg3));
-            Assert.AreEqual(fixMsg3, readFixMsg3);
+                    batch.Add(message);
+                    parser.AddToStream(CharEncoding.DefaultEncoding.GetBytes(message));
+                }
+
+                for (int i = 0; i < batchSize; i++)
+                {
+                    Assert.True(parser.ReadFixMessage(out string message));
+                    Assert.AreEqual(batch[i], message);
+                }
+
+                Assert.False(parser.ReadFixMessage(out _));
+            }
         }
 
         [Test]
         public void ReadPartialFixMessage()
         {
-            string partFixMsg1 = "8=FIX.4.2\x01" + "9=17\x01" + "35=4\x01" + "36=";
-            string partFixMsg2 = "88\x01" + "123=Y\x01" + "10=34\x01";
+            List<string> messageParts = new()
+            {
+                "abcdef8", // Junk
+                "8", // No BeginString found yet
+                "=FIX.4.2", // No BodyLength tag found yet
+                '\x01' + "9=17", // No BodyLength terminating SOH found yet
+                '\x01' + "35=4", // Message smaller than BodyLength value
+                '\x01' + "36=88\x01" + "123=Y\x01" + "10", // no CheckSum tag found yet
+                "=34", // No CheckSum terminating SOH found yet
+                "\x01"
+            };
 
-            byte[] partBytes1 = CharEncoding.DefaultEncoding.GetBytes(partFixMsg1);
-            byte[] partBytes2 = CharEncoding.DefaultEncoding.GetBytes(partFixMsg2);
+            Parser parser = new();
 
-            Parser parser = new Parser();
-            string readPartFixMsg;
+            for(int i = 0; i < messageParts.Count - 1; i++)
+            {
+                parser.AddToStream(CharEncoding.DefaultEncoding.GetBytes(messageParts[i]));
+                Assert.False(parser.ReadFixMessage(out _));
+            }
 
-            parser.AddToStream(partBytes1, partBytes1.Length);
-            Assert.False(parser.ReadFixMessage(out readPartFixMsg));
+            string expectedMessage = string.Join("", messageParts.Skip(1));
 
-            parser.AddToStream(partBytes2, partBytes2.Length);
-            Assert.True(parser.ReadFixMessage(out readPartFixMsg));
-
-            Assert.AreEqual(partFixMsg1 + partFixMsg2, readPartFixMsg);
+            parser.AddToStream(CharEncoding.DefaultEncoding.GetBytes(messageParts[^1]));
+            Assert.True(parser.ReadFixMessage(out string actualMessage));
+            Assert.AreEqual(expectedMessage, actualMessage);
         }
 
         [Test]
@@ -95,13 +126,14 @@ namespace UnitTests
             byte[] fixMsg = StrToBytes("8=TEST\x01" + "9=TEST\x01" + "35=TEST\x01" + "49=SS1\x01" + "56=RORE\x01" + "34=3\x01" + "52=20050222-16:45:53\x01" + "10=TEST\x01");
 
             Parser parser = new Parser();
-            parser.AddToStream(fixMsg, fixMsg.Length);
+            parser.AddToStream(fixMsg);
+            parser.AddToStream(StrToBytes(normalLength));
 
-            string readFixMsg;
-            Assert.Throws<QuickFix.MessageParseError>(delegate { parser.ReadFixMessage(out readFixMsg); });
+            Assert.Throws<QuickFix.MessageParseError>(delegate { parser.ReadFixMessage(out _); });
             
             // nothing thrown now because the previous call removes bad data from buffer:
-            Assert.DoesNotThrow(delegate { parser.ReadFixMessage(out readFixMsg); });
+            Assert.True(parser.ReadFixMessage(out string readFixMsg));
+            Assert.AreEqual(normalLength, readFixMsg);
         }
 
         [Test]