Use `Span` and some improved buffer management to eliminate all allocations apart from the returned string (and any buffer resizes).


Before:

|        Method |       Mean |   Error |  StdDev |   Gen0 | Allocated |
|-------------- |-----------:|--------:|--------:|-------:|----------:|
|   ReadMessage |   285.3 ns | 1.14 ns | 0.95 ns | 0.3672 |     768 B |
| Read5Messages | 1,561.8 ns | 7.86 ns | 7.35 ns | 2.9984 |    6272 B |

After:

|        Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
|-------------- |---------:|--------:|--------:|-------:|----------:|
|   ReadMessage | 106.3 ns | 0.83 ns | 0.77 ns | 0.0880 |     184 B |
| Read5Messages | 539.1 ns | 4.75 ns | 4.44 ns | 0.4396 |     920 B |

``` ini

BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.1413/22H2/2022Update/SunValley2)
AMD Ryzen 7 5700U with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.202
  [Host]     : .NET 6.0.15 (6.0.1523.11507), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.15 (6.0.1523.11507), X64 RyuJIT AVX2

```

```csharp

[MemoryDiagnoser]
public class Benchmark
{
    private readonly Parser _parser = new();
    private readonly byte[] _messageBytes = CharEncoding.DefaultEncoding.GetBytes(
        "8=FIX.4.4\u00019=61\u000135=A\u000134=1\u000149=ISLD\u000152=00000000-00:00:00.000\u000156=TW\u000198=0\u0001108=30\u000110=0\u0001");

    [Benchmark]
    public string ReadMessage()
    {
        _parser.AddToStream(_messageBytes, _messageBytes.Length);
        _parser.ReadFixMessage(out string message);
        return message;
    }
    
    [Benchmark]
    public string Read5Messages()
    {
        for (int i = 0; i < 5; i++)
        {
            _parser.AddToStream(_messageBytes, _messageBytes.Length);
        }

        string message;

        while (_parser.ReadFixMessage(out message))
        { }

        return message;
    }
}

```

</details>