@gbirchmeier 2023-07-13
This looks promising, as do a few other recent PRs by you and others.

I'm working to put out a new release right now.  It will not include this, but I'm going to see if I can improve our process internally so we can start moving on a lot of these PRs for the next release after this (and have these releases come out on a more frequent bases).